registration information verification (client and server-side) and display validation messages/info

+++ source/controllers/indexController.js
@@ -35,35 +35,66 @@ module.exports = function(app, con){
 	// post handler to store registration information in database and permanently storing profile image
 	app.post('/register', function(req, res){
 		var data = req.body;
-		// hash the password
-		bcrypt.hash(data.password, 10, function(err, hash){
-			if (err) throw err;
-			var img_dir = __dirname + "/../temp/" + data.filename;
-			var sql = "INSERT INTO users (user_id, user_pw, user_name, user_home_address, user_office_address, user_groups) VALUES ('";
-			if(data.homeaddr == "" && data.officeaddr == "")
-				sql += data.email + "', '" + hash + "', '" + data.name + "', NULL, NULL, ";
-			else if(data.homeaddr == "")
-				sql += data.email + "', '" + hash + "', '" + data.name + "', NULL, '" + data.officeaddr + "', ";
-			else if (data.officeaddr == "")
-				sql += data.email + "', '" + hash + "', '" + data.name + "', '" + data.homeaddr + "', NULL, ";
-			else
-				sql += data.email + "', '" + hash + "', '" + data.name + "', '" + data.homeaddr + "', '" + data.officeaddr + "', ";
-
-			sql += "JSON_OBJECT('type', 'user_groups', 'data', JSON_ARRAY()))"
-			// insert into table through sql query
+		// verifies the data given is valid (in case user deactivates javascript)
+		var emailRe = /[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?/ 
+		var pwRe = /^(?=.*\d)(?=.*[a-zA-Z])[0-9a-zA-Z]{7,}$/
+		if(!pwRe.test(data.password) || !emailRe.test(data.email)){
+			res.send(JSON.stringify({success:false}))
+		}else{
+			var sql = "SELECT user_id FROM users WHERE user_id=" + con.escape(data.email);
 			con.query(sql, function(err, result){
-				if (err) throw err;
-				// copy file from temp to profile_imgs folder, delete temp file
-				fs.readFile(img_dir, function(err, data){
-					if (err) throw err;
-					fs.writeFile(__dirname + "/../profile_imgs/" + result.insertId + path.extname(img_dir), data, function(err){
+				// checks if email already exists
+				if(result.length > 0)
+					res.send(JSON.stringify({success: false}));
+				else{
+					// hash the password
+					bcrypt.hash(data.password, 10, function(err, hash){
 						if (err) throw err;
-						fs.unlink(img_dir, function(err){ if (err) throw err; });
+						var img_dir = __dirname + "/../temp/" + data.filename;
+
+						// create the sql query string
+						var sql = "INSERT INTO users (user_id, user_pw, user_name, user_home_address, user_office_address, user_groups) VALUES (";
+						if(data.homeaddr == "" && data.officeaddr == "")
+							sql += `${con.escape(data.email)}, '${hash}', ${con.escape(data.name)}, NULL, NULL, `;
+						else if(data.homeaddr == "")
+							sql += `${con.escape(data.email)}, '${hash}', ${con.escape(data.name)}, NULL, ${con.escape(data.officeaddr)}, `;
+						else if (data.officeaddr == "")
+							sql += `${con.escape(data.email)}, '${hash}', ${con.escape(data.name)}, ${con.escape(data.homeaddr)}, NULL, `;
+						else
+							sql += `${con.escape(data.email)}, '${hash}', ${con.escape(data.name)}, ${con.escape(data.homeaddr)}, ${con.escape(data.officeaddr)}, `;
+						sql += "JSON_OBJECT('type', 'user_groups', 'data', JSON_ARRAY()))"
+
+						// insert new user into the table
+						con.query(sql, function(err, result){
+							if (err) throw err;
+							// copy file from temp to profile_imgs folder, and delete temp file
+							fs.readFile(img_dir, function(err, data){
+								if (err) throw err;
+								fs.writeFile(__dirname + "/../profile_imgs/" + result.insertId + path.extname(img_dir), data, function(err){
+									if (err) throw err;
+									fs.unlink(img_dir, function(err){ if (err) throw err; });
+									res.send(JSON.stringify({success: true}));
+								})
+							})
+						});
 					})
-				})
+				}
 			});
-		})
+		}
 	})
+
+	// returns true to the user if email is not already in use
+	app.post('/verify_email', function(req, res){
+		var data = req.body;
+		var sql = "SELECT user_id FROM users WHERE user_id=" + con.escape(data.email);
+		con.query(sql, function(err, result){
+			if(result.length > 0){
+				res.send(JSON.stringify({success: false}));
+			}else{
+				res.send(JSON.stringify({success: true}));
+			}
+		});
+	});
 
 	// post handler for logins: checks that userid exists and if password match
 	// returns a json to user depending on verification outcome
@@ -73,7 +104,7 @@ module.exports = function(app, con){
 			res.send(JSON.stringify({success: true, issue: false}));
 		}else{
 			var data = req.body;
-			var sql = "SELECT user_id, user_pw FROM users WHERE user_id='" + data.id +"'";
+			var sql = "SELECT user_id, user_pw FROM users WHERE user_id=" + con.escape(data.id);
 			con.query(sql, function(err, result){
 				if(result.length > 0){
 					var hashed_pw = result[0].user_pw;
